feat: move saveThemeApi function to a new utility file and update ToggleDarkMode to use it
