+++ src/shared/lib/preferences.lib.ts
@@ -28,23 +28,3 @@ export const saveThemeCookie = async (theme: ThemeMode): Promise<void> => {
     sameSite: 'strict',
   })
 }
-
-/**
- * This function calls the API to save the theme in the cookie for client side components
- * @param theme - The theme mode to save
- * @returns Promise<void>
- */
-export const saveThemeApi = async (theme: ThemeMode) => {
-  try {
-    const res = await fetch('/api/preferences/theme', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ theme }),
-    })
-    return res
-  } catch (error) {
-    console.log('error while saving theme in api', error)
-  }
-}
